+++ packages/expo-updates/ios/EXUpdates/Update/BareUpdate.swift
@@ -11,8 +11,8 @@ internal final class BareUpdate: Update {
    * application binary. They contain the minimum amount of information needed to reliably identify
    * the update and insert it into SQLite.
    */
-  public static func update(
-    withBareManifest: EXManifestsBareManifest,
+  static func update(
+    withBareManifest: BareManifest,
     config: UpdatesConfig,
     database: UpdatesDatabase?
   ) -> Update {