+++ packages/expo-updates/ios/Tests/NewUpdateSpec.swift
@@ -15,7 +15,7 @@ class NewUpdateSpec : ExpoSpec {
   override func spec() {
     describe("instantiation") {
       it("all fields") {
-        let manifest = EXManifestsNewManifest(
+        let manifest = NewManifest(
           rawManifestJSON: [
             "runtimeVersion": "1",
             "id": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -27,17 +27,8 @@ class NewUpdateSpec : ExpoSpec {
           ]
         )
 
-        let manifestHeaders = ManifestHeaders(
-          protocolVersion: nil,
-          serverDefinedHeaders: nil,
-          manifestFilters: nil,
-          manifestSignature: nil,
-          signature: nil
-        )
-
         expect(NewUpdate.update(
           withNewManifest: manifest,
-          manifestHeaders: manifestHeaders,
           extensions: [:],
           config: self.config,
           database: self.database
@@ -45,7 +36,7 @@ class NewUpdateSpec : ExpoSpec {
       }
 
       it("no runtime version") {
-        let manifest = EXManifestsNewManifest(
+        let manifest = NewManifest(
           rawManifestJSON: [
             "id": "0eef8214-4833-4089-9dff-b4138a14f196",
             "createdAt": "2020-11-11T00:17:54.797Z",
@@ -56,17 +47,8 @@ class NewUpdateSpec : ExpoSpec {
           ]
         )
 
-        let manifestHeaders = ManifestHeaders(
-          protocolVersion: nil,
-          serverDefinedHeaders: nil,
-          manifestFilters: nil,
-          manifestSignature: nil,
-          signature: nil
-        )
-
         expect(NewUpdate.update(
           withNewManifest: manifest,
-          manifestHeaders: manifestHeaders,
           extensions: [:],
           config: self.config,
           database: self.database
@@ -74,7 +56,7 @@ class NewUpdateSpec : ExpoSpec {
       }
 
       it("no id") {
-        let manifest = EXManifestsNewManifest(
+        let manifest = NewManifest(
           rawManifestJSON: [
             "runtimeVersion": "1",
             "createdAt": "2020-11-11T00:17:54.797Z",
@@ -85,17 +67,8 @@ class NewUpdateSpec : ExpoSpec {
           ]
         )
 
-        let manifestHeaders = ManifestHeaders(
-          protocolVersion: nil,
-          serverDefinedHeaders: nil,
-          manifestFilters: nil,
-          manifestSignature: nil,
-          signature: nil
-        )
-
         expect(NewUpdate.update(
           withNewManifest: manifest,
-          manifestHeaders: manifestHeaders,
           extensions: [:],
           config: self.config,
           database: self.database
@@ -103,7 +76,7 @@ class NewUpdateSpec : ExpoSpec {
       }
 
       it("no created at") {
-        let manifest = EXManifestsNewManifest(
+        let manifest = NewManifest(
           rawManifestJSON: [
             "runtimeVersion": "1",
             "id": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -114,17 +87,8 @@ class NewUpdateSpec : ExpoSpec {
           ]
         )
 
-        let manifestHeaders = ManifestHeaders(
-          protocolVersion: nil,
-          serverDefinedHeaders: nil,
-          manifestFilters: nil,
-          manifestSignature: nil,
-          signature: nil
-        )
-
         expect(NewUpdate.update(
           withNewManifest: manifest,
-          manifestHeaders: manifestHeaders,
           extensions: [:],
           config: self.config,
           database: self.database
@@ -132,7 +96,7 @@ class NewUpdateSpec : ExpoSpec {
       }
 
       it("no launch asset") {
-        let manifest = EXManifestsNewManifest(
+        let manifest = NewManifest(
           rawManifestJSON: [
             "runtimeVersion": "1",
             "id": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -140,66 +104,12 @@ class NewUpdateSpec : ExpoSpec {
           ]
         )
 
-        let manifestHeaders = ManifestHeaders(
-          protocolVersion: nil,
-          serverDefinedHeaders: nil,
-          manifestFilters: nil,
-          manifestSignature: nil,
-          signature: nil
-        )
-
         expect(NewUpdate.update(
           withNewManifest: manifest,
-          manifestHeaders: manifestHeaders,
           extensions: [:],
           config: self.config,
           database: self.database
         )).to(raiseException())
-      }
-    }
-
-    describe("dictionaryWithStructuredHeader") {
-      it("SupportedTypes") {
-        let header = "string=\"string-0000\", true=?1, false=?0, integer=47, decimal=47.5"
-        let expected: [String : Any] = [
-          "string": "string-0000",
-          "true": true,
-          "false": false,
-          "integer": 47,
-          "decimal": 47.5
-        ]
-        let actual = NewUpdate.dictionaryWithStructuredHeader(header)
-        expect(NSDictionary(dictionary: expected).isEqual(to: actual!)).to(beTrue())
-      }
-      
-      it("IgnoresOtherTypes") {
-        let header = "branch-name=\"rollout-1\", data=:w4ZibGV0w6ZydGUK:, list=(1 2)"
-        let expected: [String : Any] = [
-          "branch-name": "rollout-1"
-        ]
-        let actual = NewUpdate.dictionaryWithStructuredHeader(header)
-        expect(NSDictionary(dictionary: expected).isEqual(to: actual!)).to(beTrue())
-      }
-      
-      it("IgnoresParameters") {
-        let header = "abc=123;a=1;b=2"
-        let expected: [String : Any] = [
-          "abc": 123
-        ]
-        let actual = NewUpdate.dictionaryWithStructuredHeader(header)
-        expect(NSDictionary(dictionary: expected).isEqual(to: actual!)).to(beTrue())
-      }
-      
-      it("Empty") {
-        let header = ""
-        let expected: [String : Any] = [:]
-        let actual = NewUpdate.dictionaryWithStructuredHeader(header)
-        expect(NSDictionary(dictionary: expected).isEqual(to: actual!)).to(beTrue())
-      }
-      
-      it("ParsingError") {
-        let header = "bad dictionary"
-        expect(NewUpdate.dictionaryWithStructuredHeader(header)).to(beNil())
       }
     }
   }